Исправили отображение комментариев в миниатюрах
